moved all files related to audioplayer to domains/audio
