--- src/components/AudioPlayer/AudioProgress/index.js
@@ -1,1 +1,0 @@
-export { default } from './AudioProgress.container';
