--- src/components/AudioPlayer/Player/index.js
@@ -1,3 +1,0 @@
-import PlayerComponent from './Player.component';
-
-export default PlayerComponent;
