--- src/components/AudioPlayer/PlayerButtons/Forward/index.js
@@ -1,1 +1,0 @@
-export { default } from './ForwardButton';
