--- src/components/AudioPlayer/PlayerButtons/Play/index.js
@@ -1,1 +1,0 @@
-export { default } from './PlayButton';
