--- src/components/AudioPlayer/PlayerButtons/Repeat/index.js
@@ -1,3 +1,0 @@
-import RepeatButton from './RepeatButton';
-
-export default RepeatButton;
