--- src/components/AudioPlayer/PlayerButtons/Rewind/index.js
@@ -1,1 +1,0 @@
-export { default } from './RewindButton';
