--- src/components/AudioPlayer/PlayerButtons/Shuffle/index.js
@@ -1,3 +1,0 @@
-import ShuffleButton from './ShuffleButton';
-
-export default ShuffleButton;
