--- src/components/AudioPlayer/PlayerButtons/index.js
@@ -1,9 +1,0 @@
-import RepeatButton from './Repeat';
-import RewindButton from './Rewind';
-import PlayButton from './Play';
-import ForwardButton from './Forward';
-import ShuffleButton from './Shuffle';
-
-export {
-  RepeatButton, RewindButton, PlayButton, ForwardButton, ShuffleButton,
-};
