--- src/components/AudioPlayer/index.js
@@ -1,3 +1,0 @@
-import AudioPlayerContainer from './AudioPlayer.container';
-
-export default AudioPlayerContainer;
