+++ src/domains/audio/audioPlayer.component.js
@@ -2,11 +2,11 @@ import React from 'react';
 import propTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
 
-import PlayerComponent from './Player';
-import AudioProgressContainer from './AudioProgress';
+import PlayerComponent from './player';
+import AudioProgressContainer from './audioProgress';
 
 import './style/player.scss';
-import './AudioProgress/styles/progress.scss';
+import './audioProgress/styles/progress.scss';
 
 const useStyles = makeStyles(() => ({
   audioContainer: {
