+++ src/domains/audio/audioProgress/audioProgress.component.js
