+++ src/domains/audio/audioProgress/audioProgress.utils.js
