+++ src/domains/audio/audioProgress/index.js
@@ -1,0 +1,1 @@
+export { default } from './audioProgress.container';
