+++ src/domains/audio/audioProgress/styles/progress.scss
