+++ src/domains/audio/index.js
@@ -1,0 +1,3 @@
+import AudioPlayerContainer from './audioPlayer.container';
+
+export default AudioPlayerContainer;
