+++ src/domains/audio/player/index.js
@@ -1,0 +1,3 @@
+import PlayerComponent from './player.component';
+
+export default PlayerComponent;
