+++ src/domains/audio/playerButtons/forward/index.js
@@ -1,0 +1,1 @@
+export { default } from './forwardButton';
