+++ src/domains/audio/playerButtons/index.js
@@ -1,0 +1,9 @@
+import RepeatButton from './repeat';
+import RewindButton from './rewind';
+import PlayButton from './play';
+import ForwardButton from './forward';
+import ShuffleButton from './shuffle';
+
+export {
+  RepeatButton, RewindButton, PlayButton, ForwardButton, ShuffleButton,
+};
