+++ src/domains/audio/playerButtons/play/index.js
@@ -1,0 +1,1 @@
+export { default } from './playButton';
