+++ src/domains/audio/playerButtons/repeat/index.js
@@ -1,0 +1,3 @@
+import RepeatButton from './repeatButton';
+
+export default RepeatButton;
