+++ src/domains/audio/playerButtons/rewind/index.js
@@ -1,0 +1,1 @@
+export { default } from './rewindButton';
