+++ src/domains/audio/playerButtons/shuffle/index.js
@@ -1,0 +1,3 @@
+import ShuffleButton from './shuffleButton';
+
+export default ShuffleButton;
