+++ src/domains/audio/style/player.scss
@@ -1,6 +1,6 @@
-@import "../../../styles/colors.scss";
-@import "../../../styles/sizes.scss";
-@import "../../../styles/fonts.scss";
+@import "../../../styles/colors";
+@import "../../../styles/sizes";
+@import "../../../styles/fonts";
 .button-active {
     background-color: $green-active;
 }
