+++ src/tests/ForwardButton.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
-import ForwardButton from 'components/AudioPlayer/PlayerButtons/Forward';
+import ForwardButton from 'domains/audio/playerButtons/forward';
 
 import store from 'store';
