+++ src/tests/RewindButton.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
-import RewindButton from 'components/AudioPlayer/PlayerButtons/Rewind';
+import RewindButton from 'domains/audio/playerButtons/rewind';
 
 import store from 'store';
